blkback: if the ring slot cannot be parsed, print diagnostics
If there is a race in the frontend e.g. one thread pushes before
another thread has written a slot; or a problem with a memory
barrier then the backend can see an unparseable request. To help
diagnose this we wait a few seconds and then attempt to parse
the request again. If it can be parsed now, this suggests some
kind of race. If it still can't be parsed then this is mere
request corruption.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>